+++ zipdb.py
@@ -1,9 +1,11 @@
+from os.path import getsize
 from zipfile import ZIP_LZMA, ZipFile
 
-def compress(db: bytes, flnm: str = "exercices.db.zip") -> None:
+def compress(db: bytes, flnm: str = "exercices.db.zip") -> float:
     myzip = ZipFile(flnm, "w", compression=ZIP_LZMA, compresslevel=9)
     myzip.writestr("exercices.db", db)
     myzip.close()
+    return getsize(flnm)/len(db)
 
 def uncompress() -> tuple[bytes, bool]:
     try: